Fix PV voltage

+++ m20/Core/Inc/aprs.h
@@ -41,7 +41,7 @@ typedef struct TAPRSPacket {
 	int16_t ExtTemp;
 	uint16_t Press;
 	uint16_t BatVoltage;
-	uint8_t PvVoltage; // Payload / PV voltage
+	uint16_t PvVoltage; // Payload / PV voltage
 } APRSPacket;
 
 uint8_t encode_APRS_packet(APRSPacket Packet, char* buff);
